+++ dashboard/index.html
@@ -41,7 +41,6 @@
   <script src="js/lib/jquery-3.3.1.min.js"></script>
   <script src="js/lib/d3.min.js"></script>
   <script src="js/lib/metricsgraphics.min.js"></script>
-  <script src="js/views.js"></script>
-  <script src="js/index.js"></script>
+  <script type="module" src="js/index.js"></script>
 </body>
 </html>
